+++ src/lib/apps.ts
@@ -23,6 +23,9 @@ export class LuisApps {
           'Ocp-Apim-Subscription-Key': values.key,
         },
       };
+      if (values && values.retryDelay){
+        requestOptions["retryDelay"] = values.retryDelay
+      }
 
       const appsAsObjects = await request(requestOptions);