+++ src/lib/models.ts
@@ -35,6 +35,10 @@ export class LuisAppVersionModels {
         },
       };
 
+      if (values && values.retryDelay){
+        requestOptions["retryDelay"] = values.retryDelay
+      }
+
       const myModels = await request(requestOptions);
 
       return myModels;